+++ apps/NES/tests/test_NES_lib.v
@@ -1,25 +1,26 @@
 From elpi.apps.NES Extra Dependency "nes_synterp.elpi" as nes_synterp.
+From elpi.apps.NES Extra Dependency "nes_interp.elpi" as nes_interp.
 From elpi.apps Require Import NES.
 
 Elpi Command Make.
-#[synterp] Elpi Accumulate Db NES.db.
+#[phase="both"] Elpi Accumulate Db NES.db.
 #[synterp] Elpi Accumulate File nes_synterp.
+#[interp] Elpi Accumulate File nes_interp.
 #[synterp] Elpi Accumulate lp:{{
 
-  main-synterp [str Path] ActionsToOpen :- std.do! [
+  main [str Path] :- std.do! [
     nes.string->ns Path NS,
     nes.begin-path NS OpenNS,
-    coq.synterp-actions ActionsToOpen,
     OpenNS => nes.end-path NS _NewNS,
   ].
   main _ :- coq.error "usage: Make <DotSeparatedPath>".
 
 }}.
 #[interp] Elpi Accumulate lp:{{
-  main-interp [str _] ActionsBefore :- std.do! [
-    std.forall ActionsBefore coq.replay-synterp-action,
+  main _ :- std.do! [
+    nes.begin-path,
     coq.env.add-const "x" {{ 42 }} _ @transparent! _C,
-    coq.replay-all-missing-synterp-actions,
+    nes.end-path,
   ].
 }}.
 Elpi Typecheck.